// Demo_AntDesignOfAngular/asiatek_web/src/app/interfaces/company.ts
export interface Company {
  id: String;
  type: String;
  parentId: String;
  parentIds: String;
  name: String;
  enName: String;
  shortName: String;
  address: String;
  zipCode: String;
  fax: String;
  dutyPerson: String;
  mobile: String;
  phone: String;
  sort: String;
  createBy: String;
  createDate: String;
  updateBy: String;
  updateDate: String;
  remarks: String;
  delFlag: String;
}